Fix wrong vap_id displayed in bml_conn_map of beerocks_cli

See merge request prpl-foundation/prplmesh/prplMesh!3150

+++ controller/src/beerocks/cli/beerocks_cli_bml.h
@@ -50,7 +50,7 @@ public:
                 struct vap_t {
                     std::string ssid;
                     std::string bssid;
-                    uint8_t vap_id;
+                    int vap_id;
                     bool backhaul_vap;
                 };
